error fix

+++ tvtk/pyface/picker.py
@@ -23,7 +23,6 @@ probe for the data at that point.
 
 from traits.api import HasTraits, Trait, Long, Array, Any, Float, \
                                 Instance, Range, true, Str, false
-
 from traitsui.api import View, Group, Item, Handler
 from tvtk.api import tvtk
 from tvtk.tvtk_base import TraitRevPrefixMap, false_bool_trait
@@ -61,9 +60,9 @@ class PickedData(HasTraits):
     valid = Trait(false_bool_trait,
                   desc='specifies the validity of the pick event')
     # Id of picked point (-1 implies none was picked)
-    point_id = Int(-1, desc='the picked point ID')
+    point_id = Long(-1, desc='the picked point ID')
     # Id of picked cell (-1 implies none was picked)
-    cell_id = Int(-1, desc='the picked cell ID')
+    cell_id = Long(-1, desc='the picked cell ID')
     # World pick -- this has no ID.
     world_pick = Trait(false_bool_trait,
                        desc='specifies if the pick is a world pick.')
@@ -107,7 +106,7 @@ class PickHandler(HasTraits):
 class DefaultPickHandler(PickHandler):
     """The default handler for the picked data."""
     # Traits.
-    ID = Trait(None, None, Int, desc='the picked ID')
+    ID = Trait(None, None, Long, desc='the picked ID')
 
     coordinate = Trait(None, None, Array('d', (3,)),
                        desc='the coordinate of the picked point')
@@ -247,7 +246,6 @@ class Picker(HasTraits):
 
     # Raise the GUI on pick ?
     auto_raise = false(desc="whether to raise the picker GUI on pick")
-
 
     default_view = View(Group(Group(Item(name='pick_type'),
                                     Item(name='tolerance'), show_border=True),
